JVMS 4.3.4 signature parsing and output
The next phase of header compilation will perform type analysis, and it needs
to be able to reason about types in bytecode. This is a step towards exposing
type information in BytecodeBoundClass. Writing signatures back to strings
won't be needed until lowering is implemented, but it makes this easier to test.

MOE_MIGRATED_REVID=133627571