Fix scope handling during annotation binding
When binding annotations on a class declaration, members of that class
are not in scope (e.g. for use as class literals in annotation element
values).

Keep track of the scope that encloses class declaration in addition to
the class member scope, and use the former when binding annotations on
the class.

MOE_MIGRATED_REVID=137767162