Better type annotation disambiguation
Given the method signature `void f(java.lang.@A String x);`, `@A` is
unambiguously a type annotation. Currently this isn't handled correctly
because the bound model doesn't distinguish between `java.lang.@A
String` and `@A String`, so if `@A` can target parameters it gets added
to the parameter declaration.

To handle this better, the parser now collects all ambiguous annotations
into the declaration annotation list. In practice, the only case that
changes is method declaration annotations that occur after the formal
type parameter list, e.g. previously `@A` was classified as a type
annotation in `<T> @A void f();`.

Finally, during disambiguation annotations are now only moved from
declarations to types, and never from types to declarations.

MOE_MIGRATED_REVID=139402105