Remove redundant super type symbols from SourceTypeBoundClass
These nodes already have types for super classes and super interfaces,
and the symbol can be read from the type. Storing the symbol separately
was redundant and made it possible for the two to be out of sync.

Also fix a bug where enum superclass types (instead of superinterface
types) were set to java.lang.annotation.Enum, which was masked because
the symbol information was correct.

MOE_MIGRATED_REVID=143618845