Fix an import resolution bug
The order of the classpath and source scopes was transposed when
resolving symbols in imports. Sources should be searched first, and then
the bootclasspath, and then the regular classpath. This prevent
compilations from succeeding if the same symbol appeared on the
classpath and in source, and an import could only be resolved using the
version in source.

MOE_MIGRATED_REVID=149818281