Fix infinite recursion in canonicalize
It's possible for a class A to be both a sub-class and the enclosing
instance of another class B:

```
class Test {
  class B {}
  class A extends B {
    A.B f;
  }
}
```

During canonicalization we don't need to consider both possibilities
simultaneously. When canonicalizing B as a member of A we end up
canonicalizing A's super-type, but the enclosing instance for the second
step should be A's outer class, not A itself. This was causing
canonicalization to get stuck in an infinite loop on examples like the
one above.

MOE_MIGRATED_REVID=179932866