Report diagnostics for type conversions
When modeling primitive widening conversions, report diagnostics for
invalid conversions instead of crashing.

This rearranges the implementation to use switches on the primitive
type kinds and casts instead of double-dispatch. The new approach makes
it easier to access the position information needed to report a
diagnostic, and also makes it more explicit which constants implement
a particular conversion.

PiperOrigin-RevId: 385854947