Fix a long-standing bug with type annotation paths on raw inner class types
As described in [JVMS
4.7.20.2](https://docs.oracle.com/javase/specs/jvms/se21/html/jvms-4.html#jvms-4.7.20.2),
type annotations on nested types are located with `type_path_kind=1` type path
entries, which describe steps starting with the 'outermost part of the type for
which a type annotation is admissible'.

Turbine represents class types for inner classes as a flat list of 'simple
class types' corresponding to outer and inner classes. The 'canonicalization'
step normalizes type arguments, and also ensures there's exactly one entry for
each nested class and non-static containing class. The former doesn't apply to
raw types (we erase the entire type if any part of it is raw), but the latter
is important to ensure we compute the right number of type path steps for inner
classes.

PiperOrigin-RevId: 576882290