Test that integration test class files round-trip through turbine's class file reader
And fix a couple of small issues with inner class attributes and annotation
visibility.

Neither of these changes affect output, these attributes are handled correctly
for classes that are compiled from source.

The inner class attributes will be useful for type annotation handling (we need
to know about nested class modifiers to interpret type paths for inner
classes), and keeping track of annotation visibility is inexpensive.

This is pre-work for handling type annotations in the class file parser, it
will be helpful to test that the annotations are read in correctly
independently of adding support for associating them with types and making them
available to annotation processors.

PiperOrigin-RevId: 576884234