Attach type annotations read from bytecode to types
This change makes type annotations available to annotation processors for
types on elements that were read from bytecode.

This completes the implementation of the the turbine changes that corresponds to the javac fix made for [JDK-8225377: type annotations are not visible to javac plugins across compilation boundaries](https://bugs.openjdk.org/browse/JDK-8225377).

PiperOrigin-RevId: 578879227