Allow path to `ct.sym` to be specified via a system property
When turbine is compiled into a native image, no JAVA_HOME will be
available at runtime and the `java.home` system property won't be set.
To still provide support for `--release` in that situation, the new
`turbine.ctSymPath` system property can be set to give turbine the
required access to `ct.sym`.