Model types of elements that are missing from the classpath as ERROR types
This fixes a bug where if a class file referenced an annotation that wasn't on
the classpath, turbine would model that annotation's type as a non-ERROR type,
and then report a fatal error trying to load the missing class file. This
changes causes it to model the missing type as an ERROR type, preventing it
from incorrectly trying to access additional information about the missing
class.

PiperOrigin-RevId: 586998515