Add partial zip64 support to turbine's zip implementation
For zip64 archives that are >4GB in size, the offsets in the central directory are stored in the 'Zip64 Extended Information Extra Field'. This change doesn't add support for that to turbine, it just improves it to throw an assertion error.

Turbine generally only expects to read compile-time jars during builds, not `_deploy.jar` archives, so it would be surprising to see a >4GB archive in practice.

PiperOrigin-RevId: 621683601