Check ct.sym first before falling back to jrt
Since Turbine is used via a GraalVM native image in Bazel, jrt is not
enabled, so looking up the bootclasspath that way doesn't work.

As of Java 22, ct.sym contains signatures for all supported JDKs, so
there is no need to look in jrt.

Partial fix for https://github.com/bazelbuild/bazel/issues/21895