Return a `char` from `formatReleaseVersion`
Currently it always returns a `String` of length 1. This matches code in the JDK that is going to require updates for JDK 36, but I'm going to worry about that later.

The motivation is that this is the only place that `String#contains` is used in a context where string views could be used instead.

PiperOrigin-RevId: 690664119