Fix recovery of non-constant expressions
When parsing potentially constant expressions, turbine only needs to handle JLS 15.29 constant expressions. If it sees a non-constant expression it should ignore the expression, instead of reporting an error. This fixes a case where a diagnostic was incorrectly reported when parsing a non-constant expression that appeared as the initializer of a final field.

PiperOrigin-RevId: 755425814