Log instead of throwing errors for fields that can't be resolved
Also add a separate CANNOT_RESOLVE_FIELD diagnostic for these.

Long term turbine is move away from throwing exceptions to eagerly fail on errors towards logging and continuing, which is necessary for annotation processing to allow processors to inspect compilations with errors. But throwing is still expedient sometimes for diagnostics that are rare and hard to recover from, and that in practice will not be fixed by later rounds of annotation processing.

PiperOrigin-RevId: 831327842